Use kolla_start for managing config merges and service start
The pods now uses kolla_start with the advantage to merge the different
configs we get from host/configmap/secret which is required as
kuberentes does not support to mount configmap on top of a mounted
hostpath.

The volumes are now managed in:
1) pkg/common/volumes.go where all Volume and VolumeMounts are
   specified which are used by all pods and their controllers.
2) each controllers has its own volume definition in
   pkf/<controller>/volumes.go for the pod specific volumes.

The init container, where used, now only has the name init which
makes it more convenient to check logs using `-c init` parameter
instead of the previous long format using <pod-name>-init.

The layout of the container templates is:
templates/<pod name>/
                     kolla_config.json
                     bin/<all scripts go here>
                     config/<all service configs go here>

Pod specific changes to mention:
* nova-migration-target
  The static ssh keys are now replaced with a nova-migration-target-ssh-keys
  secret. The secret contains the private identity for the nova_migration
  user and also the pubkey in authorized_keys format. If the secret does
  not exist it gets created with the SSH keys when the secret gets created
  for the first time.

  The init container uses the script
  templates/nova-migration-target/bin/init.sh to perform required
  init actions. Right now this only sets the worker nodes internalapi
  IP in the used sshd_config.

* libvirtd
  Gets the required ssh identity from the secret created by
  nova-migration-target pod.

* nova-compute
  The init container uses the script init script to set the mandatory per
  worker node settings in nova.conf . Right now this is setting the local
  IP address at several places.

  In addition there is now a /etc/nova/nova.conf.d hostPath mount to
  /etc/nova/nova.conf.d inside the container. This directory can be used
  to manually overwite/set any host specific setting for the nova-compute
  service, like to enable only on this node debug level.

Delete following files as they are no longer/not required:
- templates/tripleo.cnf - the computes won't have DB access
- templates/migration_ssh_identity - now created and in secret
- templates/migration_authorized_keys - ow created and in secret

Note: There will be a follow up patch to move pkg/util/ssh.go to the new
common-libs repository. Also adding unit tests when moved there.