Remove environment information from CRDs and get from CM and Secret
From PoC OSP env information was passed into nova controllers via the
CR, like VIP endpoint information, rabbit url and passwords.
This removes those parameters from the CRDs and adds if needed a
CM and/or Secret CRD parameter to pass in the name of the Secret/
configMap.
The CM/Secret needs to be created manually right now. In the end those
resources get created by a controller operator.