Connect osp worker nodes to podified ctrl plane
This is an intermediate change to connect osp worker nodes to a
podified ctrl plane deployed via dev-tools. It still requires the
common-conf and secrets as before.

Also disables vif_plugging_is_fatal until multi bridge support is
also added in OCP.

To set local ip for now we pass the keystoneAPI url from the
common-config to the nova-compute and migration target, extract
the host and use ip route to get the local endpoint. In future
we'll check the routs in OCP.