Remove dependency for common-config as input
From the cnv architecture we had to create input information on
used by the nova-operator to know details on the ctl-plane. With
the ctlplane being podified in dev-tools, we can not evaluate
this data from services/routes in OCP.

This introduce an endpoint lib which is used to create a map of all
OSP endpoints identified by the app=xyz label.

Also:
* removes adding additional hosts entries as with the current
supported services this is not required. We might want to add a
general feature to add custom hosts entires for integration of
services outside of OCP which can not be resolved using DNS.

* simplifies nova.conf template for the required settings and
defaults to set.