Set a Waiting condition if secret is missing
The original goal of the commit was to test the state of the CR if the
secret is missing. However it was not possible without actually setting
a condition in the controller code when the secret is missing as the
controller set no statuses if the secret was missing.

So this patch added a new condition and a test to verify it.