[envtest]Move placement utils to a struct
This patch moves the PlacementAPI CR handling utility functions to a
commons struct to group them and to remove repeated code. I also took
the opportunity to add some documentation to the test code.

This also teased out a bug in the test code where i) the good secret is
leaked between test specs and ii) secrete creation was missing from the
spec that created keystone. This is now fixed too.