Define NovaAPI CRD
Both the API DB and the top level message bus information is passed
down as Secrets instead of letting the NovaAPI CR looking up those
information directly from a top level Nova CR. The top level Nova CR
could create such Secret and pass down the labels to the NovaAPI CR
to find it and use it. This way the NovaAPI CR is decoupled from the
Nova CR and can be instantiated independently.

The NovaAPI CR expects that the DB user already exists in the DB service
instance. NovaAPI CR will run nova-manage db sync to ensure the schema
is created / updated in the nova_api database.

In the future we might want to move the nova-manage db sync job to
another controller but right now we are not sure where.