Define NovaConductor CRD
This could represent both a super conductor or a cell conductor.

If CellName is "cell0" then this represents a super conductor. In this case
both the API and Cell level DB and message bus inputs are required.

If CellName is something else the this conductor acts as a cell
conductor and only the Cell level DB and message bus inputs are
required. The API level inputs, if provided allows, the conductor to make
upcalls (e.g. to support nova server group anti-affinity).

The NovaConductor CR assumes that both the API ,if provided, and the
Cell level DB and message bus has are accessible via the username,
password and hostname provided. NovaConductor will run
nova-manage db sync on the Cell DB but it will expect that the API level
DB already has the DB schema created before the NovaCell CR is
instantiated.