Add Nova API module
Adds an api submodule which can be used in other operators without
pulling in all the dependencies from the main go.mod file.

The core reason to have this module is to  allow other modules
(other operators) to depend on only the API structs but not
the implementation (the controller) of the operator. This will limit
the amount of transitive dependency an operator will gain by such direct
dependency on Nova.

A good existing example is between Placement and Keystone. The placement
controller implementation depends on the KeystoneService CRD and uses the
KeystoneServiceSpec struct from the keystone-operator repo. But Placement
does not need to (and therefore does not want to) depend on the
implementation of the Keystone controller. The latter has a lot longer
list of dependencies than the former. So keystone-operator was split,
the keystone-operator/api module only contains the CRD defining structs,
while the keystone-operator module contains the CRD controller
implementations.