Define NovaCell CRD
This CRD represents a nova cell.

Responsibilities
* It will create one NovaConductor CR and pass the DB and message bus
  connection information to it.

* It will create one NovaMetadata CR if the template is provided.

* It will create on NovaNoVNCProxy CR according to the template.

This CRD expects that the Cell DB and Cell message bus information it
gets points to a DB and message bus which is accessible with the provider
user and password. This CRD will pass these information forward to the
NovaConductor CRD.

This CRD expects that the API DB and API message bus it gets
optionally points to an already initialized API DB and message bus.

There are two types of NovaCell CRs
1) Normal Cells with a name other than "cell0". In this case the
   NovaConductor CR represents a cell conductor that is dedicate and
   local to this cell and never talks to any other cells. In this case
   the API DB and message bus input is optional. A deployment can have
   1..N such NovaCell CRs.

2) A single NovaCell CR with the CellName "cell0". This should always
   exist in a deployment. In this case both the Cell and the API DB and
   message bus parameters are required. And the NovaConductor CR will
   act as a super conductor for every other Cell (represented by other
   NovaCell CRs) in the deployment.