Define NovaMetadata CRD
This represents a set of nova-metadata services either running
locally in a cell or globally for all cells in the deployment.

NovaMetadata CR expects that the cell DB connection information passed
in points to a DB that is accessible and already has the Cell DB
schema populated. If provided then the same should be true for the API DB.

If the API DB connection information is provided then the nova-metadata
services will serve all cells in the deployment.