Log git build hash at operator startup
Both the docker build and the make build target is modified to get the
current commit description from git and pass it to the manager binary.
The manager will log this out at startup in a form of:

1.6635848446133595e+09	INFO	setup	Git build hash: cnosp-33-gce2645c-dirty

This will make identifying the version of the running manager a lot
easier.