Split NovaCell relates Spec structs to have Templates
We need to split what input is coming from outside (i.e. from the user)
and what input is generated by a controller for our service CRs.

This patch creates NovaConductorTemplate, NovaMetadtaTemplate,
NovaNoVNCProxyTemplate for the external input and use that in the
NovaCellSpec instead of the respective service Spec fields.

This patch also adds the NovaCellTemplate and use it in the NovaSpec
instead of the NovaCellSpec.