[NovaAPI]Initialize Status fields
Actually other than Conditions that are initialized already the rest of
the Status fields in NovaAPI does not have a different value even if it
is initialized. They are initialized to empty values and we are using `omitempty`
tags so the Status fields will be nil on the client side regardless of
initialization.