add basic precommit config
This change requires that golangci-lint is aviable to be able to run
liniting checks. these will also be run in ci and must pass to
for github checks to pass.

This change adds basic protections against ssh keys, merge conflict
lines or trailing whitespace as well as other common errors.