Just remove finalizer from KeystoneService instead of deleting
The placement controller sets OwnerReference from PlacementAPI CR to the
KeystoneService CR it creates. Also it sets a finalizer on
KeystoneService to only allow deletion of it if the owner, PlacementAPI
is deleted. K8s will do a cascade delete via the OwnerReference so the
placement controller does not need to delete the KeystoneService
directly it only need to remove its finalizer during PlacementAPI
delete.