[test]Speed up envtest
Change the default re-queue timeout of the Deployment and Job status
polling in envtest to speed up the test execution.

make test run time:
  before: 65 secs
  after: 19 secs

It could be further speed up if lib-common would take Duration instead
of an integer representing seconds as that limits us in test to the
minimum 1 second loop delay.