[CRD]Cleanup defaulting
Make sure that optional fields also have omitempty tag to get proper
defaulting.

Also gave up on defaulting the structs when the struct field is not
mentioned in the input. I think it is not possible today with
kubebuilder. Instead the input needs to contain at least {} as the field
value.