[CRD][Nova] Support different DB Services
In a real multi cell deployment it is recommended to use separate DB
services for different cells. This patch adds support to separate the
DB service for the nova_api DB from each nova cell DB (including cell0).

To support this the NovAPISpec now takes Cell0DatabaseHostname as an
input top of the existing APIDatabaseHostname field. When nova is
deployed via the Nova CR the APIDatabaseInstance name can be provided
in NovaSpec while each NovaCellTemplate in NovaSpec.cellsTemplates has
its own CellDatabaseInstance field to specific the name of the DB
service used by the given cell. The Nova controller automatically fills
the Cell0DatabaseHostname of the NovaAPISpec so that the nova-api
service has access not just the nova_api DB but also the nova_cell0 DB
as well.

Closes: #91
Closes: #89