[NovaAPI]Do not log error when Deployment not found
When the Deployment is created the first time, it is read back
immediately by lib-common. This can result in a NotFound error as the
client uses read caches. This patch changes NovaAPI not to treat this
as a real error and just return success and let the Deployment watcher
call Reconcile when the Deployment is updated.