Refactor how Specs are created
The NovaConductor case is easy as the only source of input is the
NovaCellSpec. So reduced the input parameter list to a single
NovaCellSpec and move the factory method next to the NovaCellSpec
definition.

For the NovaAPI and NovaCell cases are more complicated as they have
multiple sources of inputs. So there I opted to have the Spec struct
creation inline in the Reconcile code as otherwise I had to pass around
a bunch of parameters to the factory method.