Speed up envtest by reducing RequeueTimeout
Now that lib-common allows configuring arbitrary small RequeueTimeout
values we can take advantage of it in envtest. In envtest we simulate
success of tasks (Job, Deployment, DB) by hand and therefore we can
reduce the time they take. So we can also reduce the delay the
reconcilers take when waiting for a task to finish.

So this patch reconfigures our reconcilers when run them in envtest to
use only 10ms RequeueTimeout and adapt the tests to also look for state
changes more frequently.

This change reduces runtime of a full envtest run from 70sec to 10sec.
And the majority of that 10sec is the setup of the k8s api server
(~6secs)

To be able to do this the reconcile setup is further refactored to
expose the timeout configurability to the test.