Remove unnecessary RequeueAfter
RequeueAfter only needed when we have to wait for something that we
cannot watch for. In an other cases we can return success and let the
watch trigger a new reconciliation when our dependency is changed.

E.g. Nova does not need to wait for NovaCell(cell0).
NovaAPI does not need to wait for the Deployment.

This reduced the amount of unnecessary reconcile calls.