[envtest]Adapt to lib-common job deletion policy
lib-common changed from immediate job deletion to a TTL based job
deletion. The Jobs are deleted via setting the TTLSecondsAfterFinished
on the Job and letting the Job controller do the deletion after the TTL
expires. In a local envtest we don't have the Job controller running. So
the tests are modified to only assert that the TTLSecondsAfterFinished
is set accordingly.

Depends-on: #115
Depends-on: https://github.com/openstack-k8s-operators/lib-common/pull/78