[Nova]Finalizer to cleanup KeystoneService finalizer
The keystone-operator/api package automatically adds Nova's finalizer to
the KeystoneService created by nova. So to be able to k8s to clean up
KeystoneService when it's owner Nova is deleted Nova controller needs to
delete the this finalizer. To be able to do so Nova needs its own
finalizer to be added to itself to ensure that Nova CR is not deleted by
k8s until the Nova controller did not cleaned up. This cleanup is only
removes the finalizer from KeystoneService then removes the finalizer
from Nova. The actual CR deletion is done by k8s automatically.