[NovaAPI]Finalizer to cleanup KeystoneService finalizer
The keystone-operator/api package automatically adds Nova's finalizer to
the KeystoneEndpoint created by NovaAPI. So to be able to k8s to clean up
KeystoneEndpoint when it's owner NovaAPI is deleted NovaAPI controller needs
to delete this finalizer. To be able to do so NovaAPI needs its own
finalizer to be added to itself to ensure that NovaAPI CR is not deleted by
k8s until the NovaAPI controller cleanes up. This cleanup only
removes the finalizer from KeystoneEndpoint then removes the finalizer
from NovaAPI. The actual CR deletion is done by k8s automatically.