[sample]Make sure passwordSelectors are defaulted
Defaulting structs only work if the value '{}' is passed to the field.
If the struct field is left out from the input then the struct will get
defaulted by golang not by kubebuilder so the kubebuilder default values
will not added.

Note that if the struct field is left out from the input but the
controller reconciling the input does an Update on the instance then
that Update will apply the defaults as it will use the '{}' value for
the struct field.