fix test failures form pr 122 and 120 conflict
pr 122 intoduced usage of statefulsets for nova api
pr 120 added multicell support

pr 122 and 120 were not in git merge conflict and both passed
all test indepently. however when pr 120 was automatically
rebased when merging it did not trigger another test run...
so it merged with failing tests.

This fixes that.