[pre-commit]Speed up make target execution
By default pre-commit passes each changed file to each hook. But our
make targets ignore those files and run the target on the whole
repository on each invocation. This is a lot of repetition. So this
patch sets `pass_filenames` to `false` in our pre-commit config to
only run our make target once not on each files individually.