[CRD][Nova]Create TransportURL CRs for API and cells
This patch extends the Nova controller to create TransportURL CRs for
NovaAPI and each NovaCell template. Then passes the name of the Secret
holding the tranport URL from TransportURL.Status.SecretName to
NovaAPI.Spec.APIMessageBusSecretName and to each
NovaCell.Spec.CellMessageBusName.

A special handling is implemented to cell0 as cell0 does not have it own
dedicated message bus but it uses the API message bus.

The CRDs are adapted to the fact that the message bus is passed as the
name of a Secret and also cleanups are done to remove unnecessary
fields.

Neither NovaAPI nor NovaConductor use the passed in secret name yet.