Logging: Use ISO8601 timestamps
The default value for the zap logger is using epoc timestamps, but
that's not all that useful for us, so we are changing it to using the
ISO8601 time format instead.

The zap package from uber is already a dependency from the zap module
of the controller-runtime package we already use, so we are not adding
a new dependency.

Signed-off-by: Bogdan Dobrelya <bogdando@yahoo.com>