Remove replace line from go.mod
We checked in a replacement for openstack-operator in our main go.mod
file. We should never do that. The actual dependency has been merged in
openstack-operator so this patch removes the replace and bumps the
openstack-operator dep instead.

Such openstack-operator bump brings in bumps for keystone, lib-common,
and mariadb operators as well.