Ensure replicas is settable to 0
Each Nova service Spec has a replicas field that by definition allows
setting it to 0. But in practice a 0 input was ignored and the default
1 is used instead. The replicas field was defined as Optional, default=1
and with omitempty tag. This caused that when the nova controller called
CreateOrPatch to create or update the service CR with a 0 value the sdk
omitted the 0 value from the update as it is the empty value for the int
type used for the replicas field.

This patch removed the omitempty tag to allow propagating the `0` value.

You can read more details in https://github.com/gibizer/operator-lint/blob/dev/linters/crd/C003/README.md