Remove default from CustomServiceConfig
The default value was meaningless from actual configuration perspective
as it was just a comment. Also the field was marked with omitempty. The
default + omitempty combination made it impossible to set the
CustomServiceConfig to "" as that is the empty value and therefore
omitted by the client when serialized the request and therefore on the
server side it become defaulted.

The change in the NovaAPI test is due to the logic in the nova-api
controller that only generates the config file for the custom
configuration if that is not empty.