[EnvTest][NovaConductor]Defaulting with unstructured create
So far our envtest created our CRs by instantiating the golang struct,
defined some fields, and sent it to the API server. In
this scenario the test applies golang default values to fields without
specific value, instead of applying the default specified in the
kubebuilder annotation. With golang we cannot simulate that a field is
not provided in the input as each golang field is automatically
initialized.

To have a way to test with real defaulting behavior we need to simulate
missing fields during create.

This patch switches over to use unstructured json data to create our CRs
as there we can simply left out non required fields and let the server
side default it during json deserialization.