Remove omitempty from PasswordSelector
We choose to default the PasswordSelector and remove omitempty for
explicitness.

Note that for some strange reason it is enough to have one defaulted
field defined where the PasswordSelector struct is used and that will
trigger using the defaults defined at the field level in
PasswordSelector struct. But having no field defined does not.

This change also teased out the bug that we forgot to propagate
PasswordSelector from Nova to NovaCell to NovaConductor and to NovaAPI
this is fixed now.