Improve logging
To help troubleshooting this patch adds log when the controller returns
form reconcile either due to waiting for some external condition or
because it succeeded. All the other return point have a non nil error
causing the controller to print a stacktrace. So with this patch every
return either logs a message or throws an error.