Bumping operator deps
This patch bumps the versions of lib-common and the other operators
(keystone, mariadb, openstack) nova depends on to latest.

Note that from openstack-operator we inherit `k8s.io/apimachinery
v0.26.0` and that seems to be OK but we cannot bump the rest of k8s.io
to 0.26.0 as it breaks controller-runtime and / or controller-gen.