Fix cell iteration order
The iteration order of a map is undefined in golang. The nova controller
iterates the cellTemplates map to create NovaCell CRs. Also there is a
dependency check that a cell is not created if it needs API access and
the cell0 is not ready as that cell syncs the API DB. However the cell
template iteration order is random so in some cases cell0 is not checked
by the loop before another cell wants to check the status of the cell0.
This can lead to situation where cell0 was ready but cell1
reconciliation is not kicked off as cell0 was not yet iterated in that
loop. This caused random test failures. And this can cause that cell1
status is flipping between Ready and Not Ready.

This patch makes sure that the cells are iterated in an order where
cell0 is always handled first.