Generate manifests before test run
The envtest uses our CRD definitions during test so we need to run `make
manifests` and `make generate` before the test run to always test with
an up to date CRD yaml and golang definition.

The `make test` target generates manifests today but the `make gotest`
target doesn't. The latter is used in the prow CI job[1]. So this patch
make `gotest` an alias for the `test` target.

Also this patch switches the github action based test run to use `make
gotest` instead of indirectly calling `go test`

[1] https://github.com/openshift/release/blob/fe0b5ddd5c96f053a654883ae33ec05933ad6c5c/ci-operator/config/openstack-k8s-operators/nova-operator/openstack-k8s-operators-nova-operator-master.yaml#L31