[CI] workflow for operator image and bundle build
This patch adds a github workflow to build the nova-operator image,
and bundle for each PR to see if the PR does not break the build.
This job will not publish the build artifacts. Another build will run
after the PR is merged to build the official artifacts and publish them.