add example ansible playbooks
This change adds 3 playbooks to show how to run run playbooks
via the AnsibleEE operator.

The first example "ansibleee-playbook-local.yaml" defines
the inventory and playbook inline in the AnsibleEE CR.
This execute directly against the AnsibleEE operator container.

The second exmaple "ansibleee-external-inventory.yaml" uses the extra
mounts functionality to inject an inventory and ssh keys into the
AnsibleEE operator to run the inline playbook against a remote host.

The final example "lsblk.yaml" build on the second to also inject
the playbook as a projection from a config map.
Taking this approch we could package the playbooks in the nova-operator
bundel and have the operator import them into a config map which can
be attached to the AnsibleEE instance.