add compute node ansible support
This is the inital commit in the support for external compute
hosts managed via ansible. This is created from my poc in PR #233
by taking the final working version as a base.

Following commits will refactor this to be more generic and
relocate the playbooks and configs so that they can be integrated
with the golang operator.