Make sure renovate runs make manifests generate
After a dependency bump the yaml and go code generators need to be run
to pick up changes. Now that we have a hosted renovate runner we can
configure a postUpgradeTask to do so.

Note that the fileFilters defining what changes are added to the commit
after the commands are run.