update playbooks and add remaining libvirt config files
This change updates the config map generation to include all libvirt
and nova compute config files and contianer files.

This change updates the playbooks to copy the container files
using the new names used in the config map.

As we group the container files per folder its technialy possible you
could have a name collistion between directores when the files are
flatened into the config map. To prevent this the path fo the templeate
releitive to the template root dir is used as the key with path seperators
replaced to __ to workaround the fact that configmaps do not allow /
in key names.