Reproduce that Ready is not reset during reconfiguration
If NovaAPI reaches the Ready True state and then reconfigured then the
Ready condition is not reset. So if a reconfiguration step fails then
the Status.Conditions inconsistently shows Ready True + a subcondition
in False state.

This patch adds a test scenario that shows the failure for
NovaConductor, NovaScheduler, NovaCell, and Nova.