Use the internal Keystone endpoint in the configs
We use the public endpoint so far but only the internal one is exposed
to the external dataplane nodes so for consistency reason we change all
the nova services to use that. Note that the one in the NovaCell.Spec
will be used by the NovaExternalCompute to generate the compute config.