Fix hard-coded playbooks path
When running with `make run` the hard-coded path to the playbooks is
wrong and makes the operator unable to find them. Running with the
OPERATOR_PLAYBOOKS variable set appropriately *does* work, and changing
this path makes it work for me without as well. The current value is
set so it works for the tests, but that doesn't make much sense since
they are run from the atypical location of test/functional. Correct the
default so it works for the more usual case of `make run`.