Move OPERATOR_ variables to Makefile for tests
The functional tests are run from a different place than the operator
itself, which means the hard-coded path in the operator code does not
work for the tests. Instead of hard-coding the variable override into
the tests themselves, we should set it in the Makefile, which is what
actually controls the location from which they are run.