Convert conductor to new config gen
This change removes the use of crudini in nova and finishes
the conversion of config generation to go templates.

As part of this work the config generation is update to only include
the api db info if a api db is spcified in the CR and the secret
validation is not dynmic.

we wil now ensure all field that are used are asserted as present
in the secrete before we use them and dynmically include the api
db secret when correct to do so.