Fix NovaAPIMQReadyCondition error reporting
To report an error in  NovaAPIMQReadyCondition nova used the error
from the API DB creation instead of the error from the API MQ creation.
This could lead to a missleading error message if the API DB creation
was also failed. Or could lead to operator crash if the API DB error was
nil.