Do the cell registration from the nova controller
So far the cell registration happened in the DB sync job run by the
NovaConductor controller. However for cells without API DB access this
solution does not work. So this patch separates the DB sync and the DB
registration tasks to separate Jobs. The DB sync is still run by the
NovaConductor controller. But the DB registration is run the Nova
controller after a given NovaCell becomes Ready.

Also the Nova controller now records in its Status which cell is mapped
and with what configuration. So that if the configuration of a cell
changes and therefore the cell mapping needs an update then Nova
controller will detect that and re-run the Job to update the Nova API
DB.

Closes: https://issues.redhat.com/browse/OSPRH-100
Closes: #147
Closes: #276