Make top level deployment depend on registered cells
The top level services needs to be restarted when a new cell is
registered or an existing cell registration is updated in the nova_api
database.

To automate the service restart a new RegisteredCells field is added
to the top level service CRDs. This field will be updated by
the Nova CR controller based on the cells it maps in the nova_api
database.

Also this patch includes the hash of the RegisteredCells field
of the top level services into the environment of the StatefulSet of
those services. The StatufulSet will automatically restart the Pods
when the hash changes.

Closes: https://issues.redhat.com/browse/OSPRH-107