Add RegisteredCells to the NovaStatus
This field will store the cells and their config hash nova-controller
already registered in the nova_api database. Later this data will be
distributed from Nova.Status to each top level service Spec so that the
service can implement an automatic Pod restart if a cell is registered
or its config is updated.