[envtest]Remove duplicated NovaScheduler instance
The NovaScheduler controller test created a NovaScheduler instance
before each test. However there are test cases later in the file that
also creates its own NovaScheduler instance with specific configuration.
So in those tests there are two NovaScheduler instances being
reconciled. This makes troubleshooting the test complex. So this patch
separates the tests with the global NovaScheduler to a deparate Describe
section. So that the rest of the tests can  create its own NovaScheduler
with the needed configuration.