Make ServiceAccount Required in sub CRs
It was only defaulted in
https://github.com/openstack-k8s-operators/nova-operator/pull/347
to avoid breaking the test. But semantically we want them to be
Required. So this patch removes the default value from the field
and update the test to provider ServiceAccount value to sub CRs.
Also updates our samples with serviceAccount fields to keep them valid.