Fix cell0 access while cell0 waits for DB
There was a bug in the nova-controller in the ensureCell loop. If the
cell0 was still waiting for the DB to be created but cell1 had all the
prerequisites are done (DB, MQ) then the code tried to look up cell0
which was not in the internal cell map causing panic.