Watch Spec.Secret
If the content of the Secret pointed by the Nova*.Spec.Secret fields
changes then the given CR will be reconciled. If one of the relevant
passwords in that Secret is changed then the affected nova service
configs are regenerated and the CONFIG_HASH in the statefulsets are
updated to trigger service restart.

Closes: https://issues.redhat.com/browse/OSPRH-215