Sync the OWNERS file with the other repositories
All the other repositories has the structure that allows approving
patches for CI folks and the podified core team alongside with the repo
specific approver group.

For some reason nova-operator repo was not synced to this structure when
the rest of the repo was synced. But there was a discussion in slack[1]
over it. There was some objection over this in slack. Lets try to
conclude this now.

[1] https://redhat-internal.slack.com/archives/CQXJFGMK6/p1680527086750569