Added playbook to install meta operator
https://github.com/openstack-k8s-operators/ci-framework/pull/201 changes
the parent from base-crc to base-simple-crc.

base-simple-crc job will provide an crc based openshift cluster. It
lacks the meta operator installation bits which breaks the existing job.

https://github.com/openstack-k8s-operators/ci-framework/pull/245 refines
the cifmw-base-crc-openstack job and moves the ci-framework pre
playbooks in pre-run and collect-logs in post-run playbook. It will help
to drop the existing collect-logs.yaml playbook.

In order to fix the job, It uses install_yamls_make role to install
meta operator.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/272

Signed-off-by: Chandan Kumar <raukadah@gmail.com>